pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "deft"
version = "0.1.0"
requires-python = ">=3.10,<3.11"
dependencies = [
    "accelerate>=1.12.0",
    "aiohttp==3.11.18",
    "captum==0.8.0",
    "datasets==4.1.1",
    "dill==0.4.0",
    "genomic-benchmarks==0.0.9",
    "h5py==3.14.0",
    "hydra-core==1.3.2",
    "loky==3.5.6",
    "matplotlib==3.8.4",
    "numpy==1.26.4",
    "omegaconf==2.3.0",
    "openai==0.28.1",
    "pandas==2.3.3",
    "peft==0.17.1",
    "pytorch-lightning==2.5.5",
    "pyyaml>=6.0.3",
    "radon>=6.0.1",
    "reportlab>=4.4.10",
    "scienceplots>=2.2.0",
    "scikit-learn==1.6.1",
    "scipy==1.15.3",
    "seaborn>=0.13.2",
    "together==1.5.21",
    "torch==2.8.0",
    "torchmetrics==1.2.1",
    "tqdm==4.67.1",
    "transformers==4.57.0",
    "xgboost==3.0.2",
]

[tool.setuptools.packages.find]
include = ["src*"] # finds 'src' and 'src.*'


